addDiagnosePlpToDatabase {PatientLevelPrediction}	R Documentation

Insert a diagnostic result into a PLP result schema database

Description

This function inserts a diagnostic result into the result schema

Usage

addDiagnosePlpToDatabase(
  diagnosePlp,
  connectionDetails,
  databaseSchemaSettings,
  cohortDefinitions,
  databaseList = NULL,
  overWriteIfExists = T
)

Arguments

diagnosePlp
An object of class diagnosePlp

connectionDetails
A connection details created by using the function createConnectionDetails in the DatabaseConnector package.

databaseSchemaSettings
A object created by createDatabaseSchemaSettings with all the settings specifying the result tables

cohortDefinitions
A set of one or more cohorts extracted using ROhdsiWebApi::exportCohortDefinitionSet()

databaseList
(Optional) If you wish to overwrite the settings in the plp object use createdatabaseList to specify the databases

overWriteIfExists
(default: T) Whether to delete existing results and overwrite them

Details

This function can be used to upload a diagnostic result into a database

Value

Returns NULL but uploads the diagnostic into the database schema specified in databaseSchemaSettings

[Package PatientLevelPrediction version 6.3.1 Index]