addMultipleDiagnosePlpToDatabase {PatientLevelPrediction}	R Documentation

Insert mutliple diagnosePlp results saved to a directory into a PLP result schema database

Description

This function inserts diagnosePlp results into the result schema

Usage

addMultipleDiagnosePlpToDatabase(
  connectionDetails,
  databaseSchemaSettings,
  cohortDefinitions,
  databaseList = NULL,
  resultLocation
)

Arguments

connectionDetails
A connection details created by using the function createConnectionDetails in the DatabaseConnector package.

databaseSchemaSettings
A object created by createDatabaseSchemaSettings with all the settings specifying the result tables

cohortDefinitions
(list) A list of cohortDefinitions (each list must contain: name, id)

databaseList
(Optional) ...

resultLocation
The location of the diagnostic results

Details

This function can be used to upload diagnosePlp results into a database

Value

Returns NULL but uploads multiple diagnosePlp results into the database schema specified in databaseSchemaSettings

[Package PatientLevelPrediction version 6.3.1 Index]