addRunPlpToDatabase {PatientLevelPrediction}	R Documentation

Function to add the run plp (development or validation) to database

Description

This function adds a runPlp or external validation result into a database

Usage

addRunPlpToDatabase(
  runPlp,
  connectionDetails,
  databaseSchemaSettings,
  cohortDefinitions,
  modelSaveLocation,
  databaseList = NULL
)

Arguments

runPlp
An object of class runPlp or class externalValidatePlp

connectionDetails
A connection details created by using the function createConnectionDetails in the DatabaseConnector package.

databaseSchemaSettings
A object created by createDatabaseSchemaSettings with all the settings specifying the result tables

cohortDefinitions
A set of one or more cohorts extracted using ROhdsiWebApi::exportCohortDefinitionSet()

modelSaveLocation
The location of the directory that models will be saved to

databaseList
(Optional) If you want to change the database name then used createDatabaseList to specify the database settings but use the same cdmDatabaseId was model development/validation

Details

This function is used when inserting results into the PatientLevelPrediction database results schema

Value

Returns a data.frame with the database details

[Package PatientLevelPrediction version 6.3.1 Index]