combinePlpModelSettings {PatientLevelPrediction}	R Documentation

combine two objects specifying multiple Plp model settings

Description

combine two objects specifying multiple Plp model settings

Usage

combinePlpModelSettings(plpModelSetting1, plpModelSetting2)

Arguments

plpModelSetting1
A combination of model, covariate and population settings

plpModelSetting2
A combination of model, covariate and population settings

Details

Takes two output of running createPlpModelSettings() and combined them

Value

A list containing a dataframe settingLookupTable containing all the model, covariate and popualtion combination details, a list models containing all the model settings, a list covariateSettings containing all the covariate settings and a list populationSettings containing all the population settings.

[Package PatientLevelPrediction version 4.3.10 Index]