computeAucFromDataFrames {PatientLevelPrediction}	R Documentation

Compute the area under the ROC curve

Description

Compute the area under the ROC curve

Usage

computeAucFromDataFrames(
  prediction,
  status,
  time = NULL,
  confidenceInterval = FALSE,
  timePoint,
  modelType = "logistic"
)

Arguments

prediction
A vector with the predicted hazard rate.

status
A vector with the status of 1 (event) or 0 (no event).

time
Only for survival models: a vector with the time to event or censor (which ever comes first).

confidenceInterval
Should 95 percebt confidence intervals be computed?

timePoint
Only for survival models: time point when the AUC should be evaluated

modelType
Type of model. Currently supported are "logistic" and "survival".

Details

Computes the area under the ROC curve for the predicted probabilities, given the true observed outcomes.

[Package PatientLevelPrediction version 4.3.10 Index]