createDatabaseList {PatientLevelPrediction}	R Documentation

Create a list with the database details and database meta data entries

Description

This function creates a list with the database details and database meta data entries used in the study

Usage

createDatabaseList(cdmDatabaseSchemas, cdmDatabaseNames, databaseRefIds = NULL)

Arguments

cdmDatabaseSchemas
(string vector) A vector of the cdmDatabaseSchemas used in the study - if the schemas are not unique per database please also specify databaseRefId

cdmDatabaseNames
Sharable names for the databases

databaseRefIds
(string vector) Unique database identifiers - what you specified as cdmDatabaseId in PatientLevelPrediction::createDatabaseDetails() when developing the models

Details

This function is used when inserting database details into the PatientLevelPrediction database results schema

Value

Returns a data.frame with the database details

[Package PatientLevelPrediction version 6.3.1 Index]