createDefaultExecuteSettings {PatientLevelPrediction}	R Documentation

Creates default list of settings specifying what parts of runPlp to execute

Description

Creates default list of settings specifying what parts of runPlp to execute

Usage

createDefaultExecuteSettings()

Details

runs split, preprocess, model development and covariate summary

Value

list with TRUE for split, preprocess, model development and covariate summary

[Package PatientLevelPrediction version 5.0.5 Index]