createExecuteSettings {PatientLevelPrediction}	R Documentation

Creates list of settings specifying what parts of runPlp to execute

Description

Creates list of settings specifying what parts of runPlp to execute

Usage

createExecuteSettings(
  runSplitData = F,
  runSampleData = F,
  runfeatureEngineering = F,
  runPreprocessData = F,
  runModelDevelopment = F,
  runCovariateSummary = F
)

Arguments

runSplitData
TRUE or FALSE whether to split data into train/test

runSampleData
TRUE or FALSE whether to over or under sample

runfeatureEngineering
TRUE or FALSE whether to do feature engineering

runPreprocessData
TRUE or FALSE whether to do preprocessing

runModelDevelopment
TRUE or FALSE whether to develop the model

runCovariateSummary
TRUE or FALSE whether to create covariate summary

Details

define what parts of runPlp to execute

Value

list with TRUE/FALSE for each part of runPlp

[Package PatientLevelPrediction version 5.0.5 Index]