createFeatureEngineeringSettings {PatientLevelPrediction}	R Documentation

Create the settings for defining any feature engineering that will be done

Description

Create the settings for defining any feature engineering that will be done

Usage

createFeatureEngineeringSettings(type = "none")

Arguments

type
(character) Choice of:

'none' No feature engineering - this is the default

Details

Returns an object of class featureEngineeringSettings that specifies the sampling function that will be called and the settings

Value

An object of class featureEngineeringSettings

[Package PatientLevelPrediction version 5.0.5 Index]