createLogSettings {PatientLevelPrediction}	R Documentation

Create the settings for logging the progression of the analysis

Description

Create the settings for logging the progression of the analysis

Usage

createLogSettings(verbosity = "DEBUG", timeStamp = T, logName = "runPlp Log")

Arguments

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

timeStamp
If TRUE a timestamp will be added to each logging statement. Automatically switched on for TRACE level.

logName
A string reference for the logger

Details

Returns an object of class logSettings that specifies the logger settings

Value

An object of class logSettings

[Package PatientLevelPrediction version 5.0.5 Index]