Convert logistic regression model to sql code...

Convert logistic regression model to sql code...

createLrSql(
  models,
  modelNames,
  covariateConstructionName = "prediction",
  modelTable = "#model_table",
  analysisId = 111,
  e = environment(),
  databaseOutput = NULL
)

Arguments

models
A trianed plp model.

modelNames
A name used in the covariate function names (no spaces)

covariateConstructionName
the name used for the create covariate function

modelTable
The temporary table name storing the model details

analysisId
The covariate analysis_id

e
The environment to output the covariate setting functions to

databaseOutput
If you want to output to go inot a cohort table add the "database.schema.tablename" here

Details

This function is used to create custom covariates for a logistic regression model (currently only supports, demographics/conditions/drug/procedures/observations and measurement concepts)