createModelDesign {PatientLevelPrediction}	R Documentation

Specify settings for deceloping a single model

Description

Specify settings for deceloping a single model

Usage

createModelDesign(
  targetId,
  outcomeId,
  restrictPlpDataSettings = createRestrictPlpDataSettings(),
  populationSettings = createStudyPopulationSettings(),
  covariateSettings = FeatureExtraction::createDefaultCovariateSettings(),
  featureEngineeringSettings = NULL,
  sampleSettings = NULL,
  preprocessSettings = NULL,
  modelSettings = NULL,
  runCovariateSummary = T
)

Arguments

targetId
The id of the target cohort that will be used for data extraction (e.g., the ATLAS id)

outcomeId
The id of the outcome that will be used for data extraction (e.g., the ATLAS id)

restrictPlpDataSettings
The settings specifying the extra restriction settings when extracting the data created using createRestrictPlpDataSettings().

populationSettings
The population settings specified by createStudyPopulationSettings()

covariateSettings
The covariate settings, this can be a list or a single 'covariateSetting' object.

featureEngineeringSettings
Either NULL or an object of class featureEngineeringSettings specifying any feature engineering used during model development

sampleSettings
Either NULL or an object of class sampleSettings with the over/under sampling settings used for model development

preprocessSettings
Either NULL or an object of class preprocessSettings created using createPreprocessingSettings()

modelSettings
The model settings such as setLassoLogisticRegression()

runCovariateSummary
Whether to run the covariateSummary

Details

This specifies a single analysis for developing as single model

Value

A list with analysis settings used to develop a single prediction model

[Package PatientLevelPrediction version 5.0.5 Index]