createPlpJournalDocument {PatientLevelPrediction}	R Documentation

createPlpJournalDocument

Description

Creates a template for a prediction journal paper with the characteristics/results filled in

Usage

createPlpJournalDocument(
  plpResult = NULL,
  plpValidation = NULL,
  plpData = NULL,
  targetName = "<target population>",
  outcomeName = "<outcome>",
  table1 = F,
  connectionDetails = NULL,
  includeTrain = FALSE,
  includeTest = TRUE,
  includePredictionPicture = TRUE,
  includeAttritionPlot = TRUE,
  outputLocation = file.path(getwd(), "plp_journal_document.docx"),
  save = T
)

Arguments

plpResult
An object of type plpResult returned by running runPlp()

plpValidation
An object of type validatePlp returned by running externalValidatePlp()

plpData
The plpData

targetName
A string with the target description name

outcomeName
A string with the outcome description name

table1
Whether to include table1 (characteristics)

connectionDetails
The connection required to calcualte the characteristics

includeTrain
Whether to include the train set performance

includeTest
Whether to include the test set performance

includePredictionPicture
Whether to include a picture detailing the prediction problem

includeAttritionPlot
Whether to include the attriction plot

outputLocation
The location to write the document to

save
If false this fucntion returns the document and does not save to outputLocation

Details

The function creates a word document containing the analysis details, data summary and prediction model results.

Value

A work document containing the selected outputs within the user's directory at location specified in outputLocation

[Package PatientLevelPrediction version 4.3.10 Index]