createPlpModelSettings {PatientLevelPrediction}	R Documentation

create a an object specifying the multiple Plp model settings

Description

create a an object specifying the multiple Plp model settings

Usage

createPlpModelSettings(modelList, covariateSettingList, populationSettingList)

Arguments

modelList
A list of model settings

covariateSettingList
A list of covariate settings

populationSettingList
A list of population settings

Details

Takes a list of models, covariates, population and returns the cartesian product combining all settings.

Value

A list containing a dataframe settingLookupTable containing all the model, covariate and popualtion combination details, a list models containing all the model settings, a list covariateSettings containing all the covariate settings and a list populationSettings containing all the population settings.

[Package PatientLevelPrediction version 4.3.10 Index]