createPlpReport

Creates a word document report of the prediction

createPlpReport(
  plpResult = NULL,
  plpValidation = NULL,
  plpData = NULL,
  targetName = "<target population>",
  outcomeName = "<outcome>",
  targetDefinition = NULL,
  outcomeDefinition = NULL,
  outputLocation = file.path(getwd(), "plp_report.docx"),
  save = T
)

Arguments

plpResult
An object of type plpResult returned by running runPlp()

plpValidation
An object of type validatePlp returned by running externalValidatePlp()

plpData
The plpData

targetName
A string with the target description name

outcomeName
A string with the outcome description name

targetDefinition
The cohort details

outcomeDefinition
The cohort details

outputLocation
The location to write the document to

save
If false the output of the function of the function is the document rather than creating the document in outputLocation

Value

A work document containing the selected outputs within the user's directory at location specified in outputLocation

Details

The function creates a word document containing the analysis details, data summary and prediction model results.