createPreprocessSettings {PatientLevelPrediction}	R Documentation

Create the settings for preprocessing the trainData using .

Description

Create the settings for preprocessing the trainData using .

Usage

createPreprocessSettings(
  minFraction = 0.001,
  normalize = TRUE,
  removeRedundancy = TRUE
)

Arguments

minFraction
The minimum fraction of target population who must have a covariate for it to be included in the model training

normalize
Whether to normalise the covariates before training (Default: TRUE)

removeRedundancy
Whether to remove redundant features (Default: TRUE)

Details

Returns an object of class preprocessingSettings that specifies how to preprocess the training data

Value

An object of class preprocessingSettings

[Package PatientLevelPrediction version 5.0.5 Index]