createRandomForestFeatureSelection {PatientLevelPrediction}	R Documentation

Create the settings for random foreat based feature selection

Description

Create the settings for random foreat based feature selection

Usage

createRandomForestFeatureSelection(ntrees = 2000, maxDepth = 17)

Arguments

ntrees
number of tree in forest

maxDepth
MAx depth of each tree

Details

Returns an object of class featureEngineeringSettings that specifies the sampling function that will be called and the settings

Value

An object of class featureEngineeringSettings

[Package PatientLevelPrediction version 5.0.5 Index]