createUnivariateFeatureSelection {PatientLevelPrediction}	R Documentation

Create the settings for defining any feature selection that will be done

Description

Create the settings for defining any feature selection that will be done

Usage

createUnivariateFeatureSelection(k = 100)

Arguments

k
This function returns the K features most associated (univariately) to the outcome

Details

Returns an object of class featureEngineeringSettings that specifies the sampling function that will be called and the settings

Value

An object of class featureEngineeringSettings

[Package PatientLevelPrediction version 5.0.5 Index]