createValidationSettings {PatientLevelPrediction}	R Documentation

createValidationSettings define optional settings for performing external validation

Description

This function creates the settings required by externalValidatePlp

Usage

createValidationSettings(recalibrate = NULL, runCovariateSummary = T)

Arguments

recalibrate
A vector of characters specifying the recalibration method to apply

runCovariateSummary
Whether to run the covariate summary for the validation data

Details

Users need to specify whether they want to sample or recalibate when performing external validation

Value

A setting object of class validationSettings containing a list of settings for externalValidatePlp

[Package PatientLevelPrediction version 5.0.5 Index]