drawAttritionDiagramPlp {PatientLevelPrediction}	R Documentation

Draw the attrition diagram

Description

drawAttritionDiagramPlp draws the attition diagram, showing how many people were excluded from the study population, and for what reasons.

Usage

drawAttritionDiagramPlp(
  attrition,
  targetLabel = "Target Population",
  outcomeLabel = "Outcome Count",
  fileName = NULL
)

Arguments

attrition
The table of attrition details return from the population attr(popualtion, 'metaData')$attrition

targetLabel
A label to us for the treated cohort.

outcomeLabel
A label to us for the comparator cohort.

fileName
Name of the file where the plot should be saved, for example 'plot.png'. See the function ggsave in the ggplot2 package for supported file formats.

Value

A ggplot object. Use the ggsave function to save to file in a different format.

[Package PatientLevelPrediction version 4.3.10 Index]