evaluatePlp {PatientLevelPrediction}	R Documentation

evaluatePlp

Description

Evaluates the performance of the patient level prediction model

Usage

evaluatePlp(prediction, typeColumn = "evaluationType")

Arguments

prediction
The patient level prediction model's prediction

typeColumn
The column name in the prediction object that is used to stratify the evaluation

Details

The function calculates various metrics to measure the performance of the model

Value

A list containing the performance values

[Package PatientLevelPrediction version 5.0.5 Index]