externalValidatePlp {PatientLevelPrediction}	R Documentation

externalValidatePlp - Validate a model on new databases

Description

This function extracts data using a user specified connection and cdm_schema, applied the model and then calcualtes the performance

Usage

externalValidatePlp(
  plpResult,
  connectionDetails,
  validationSchemaTarget,
  validationSchemaOutcome,
  validationSchemaCdm,
  databaseNames,
  validationTableTarget = "cohort",
  validationTableOutcome = "cohort",
  validationIdTarget = NULL,
  validationIdOutcome = NULL,
  oracleTempSchema = NULL,
  verbosity = "INFO",
  keepPrediction = F,
  recalibrate = NULL,
  sampleSize = NULL,
  outputFolder
)

Arguments

plpResult
The object returned by runPlp() containing the trained model

connectionDetails
The connection details for extracting the new data

validationSchemaTarget
A string or vector of strings specifying the database containing the target cohorts

validationSchemaOutcome
A string or vector of strings specifying the database containing the outcome cohorts

validationSchemaCdm
A string or vector of strings specifying the database containing the cdm

databaseNames
A string of vector of strings specifying sharing friendly database names corresponding to validationSchemaCdm

validationTableTarget
A string or vector of strings specifying the table containing the target cohorts

validationTableOutcome
A string or vector of strings specifying the table containing the outcome cohorts

validationIdTarget
An iteger specifying the cohort id for the target cohort

validationIdOutcome
An iteger specifying the cohort id for the outcome cohort

oracleTempSchema
The temp oracle schema requires read/write

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

keepPrediction
Whether to keep the predicitons for the new data

recalibrate
A vector of characters specifying the recalibration method to apply

sampleSize
If not NULL, the number of people to sample from the target cohort

outputFolder
If you want to save the results enter the directory to save here

Details

Users need to input a trained model (the output of runPlp()) and new database connections. The function will return a list of length equal to the number of cdm_schemas input with the performance on the new data

Value

A list containing the performance for each validation_schema

[Package PatientLevelPrediction version 4.3.10 Index]