getCohortCovariateData {PatientLevelPrediction}	R Documentation

Extracts covariates based on cohorts

Description

Extracts covariates based on cohorts

Usage

getCohortCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cdmVersion = "5",
  cohortTable = "#cohort_person",
  rowIdField = "row_id",
  aggregated,
  cohortId,
  covariateSettings
)

Arguments

connection
The database connection

oracleTempSchema
The temp schema if using oracle

cdmDatabaseSchema
The schema of the OMOP CDM data

cdmVersion
version of the OMOP CDM data

cohortTable
the table name that contains the target population cohort

rowIdField
string representing the unique identifier in the target population cohort

aggregated
whether the covariate should be aggregated

cohortId
cohort id for the target population cohort

covariateSettings
settings for the covariate cohorts and time periods

Details

The user specifies a cohort and time period and then a covariate is constructed whether they are in the cohort during the time periods relative to target population cohort index

Value

The models will now be in the package

[Package PatientLevelPrediction version 5.0.5 Index]