ici {PatientLevelPrediction}	R Documentation

Calculate the Integrated Calibration Information from Austin and Steyerberg https://onlinelibrary.wiley.com/doi/full/10.1002/sim.8281

Description

Calculate the Integrated Calibration Information from Austin and Steyerberg https://onlinelibrary.wiley.com/doi/full/10.1002/sim.8281

Usage

ici(prediction)

Arguments

prediction
the prediction object found in the plpResult object

Details

Calculate the Integrated Calibration Information

Value

Integrated Calibration Information

[Package PatientLevelPrediction version 5.0.5 Index]