insertModelDesignInDatabase {PatientLevelPrediction}	R Documentation

Insert a model design into a PLP result schema database

Description

This function inserts a model design and all the settings into the result schema

Usage

insertModelDesignInDatabase(
  object,
  conn,
  databaseSchemaSettings,
  cohortDefinitions
)

Arguments

object
An object of class modelDesign, runPlp or externalValidatePlp

conn
A connection to a database created by using the function connect in the DatabaseConnector package.

databaseSchemaSettings
A object created by createDatabaseSchemaSettings with all the settings specifying the result tables

cohortDefinitions
A set of one or more cohorts extracted using ROhdsiWebApi::exportCohortDefinitionSet()

Details

This function can be used to upload a model design into a database

Value

Returns NULL but uploads the model design into the database schema specified in databaseSchemaSettings

[Package PatientLevelPrediction version 6.3.1 Index]