insertResultsToSqlite {PatientLevelPrediction}	R Documentation

Create sqlite database with the results

Description

This function create an sqlite database with the PLP result schema and inserts all results

Usage

insertResultsToSqlite(
  resultLocation,
  cohortDefinitions,
  databaseList = NULL,
  sqliteLocation = file.path(resultLocation, "sqlite")
)

Arguments

resultLocation
(string) location of directory where the main package results were saved

cohortDefinitions
A set of one or more cohorts extracted using ROhdsiWebApi::exportCohortDefinitionSet()

databaseList
A list created by createDatabaseList to specify the databases

sqliteLocation
(string) location of directory where the sqlite database will be saved

Details

This function can be used upload PatientLevelPrediction results into an sqlite database

Value

Returns the location of the sqlite database file

[Package PatientLevelPrediction version 6.3.1 Index]