launchDiagnosticsExplorer {PatientLevelPrediction}	R Documentation

Launch the Diagnostics Explorer Shiny app

Description

Launch the Diagnostics Explorer Shiny app

Usage

launchDiagnosticsExplorer(dataFolder, launch.browser = FALSE)

Arguments

dataFolder
A folder where the exported zip files with the results are stored. Zip files containing results from multiple databases can be placed in the same folder.

launch.browser
Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

Details

Launches a Shiny app that allows the user to explore the diagnostics

[Package PatientLevelPrediction version 5.0.5 Index]