loadEnsemblePlpModel {PatientLevelPrediction}	R Documentation

loads the Ensmeble plp model and return a model list

Description

loads the Ensmeble plp model and return a model list

Usage

loadEnsemblePlpModel(dirPath)

Arguments

dirPath
The location of the model

Details

Loads a plp model list that was saved using savePlpModel()

[Package PatientLevelPrediction version 4.3.10 Index]