loadPlpData {PatientLevelPrediction}	R Documentation

Load the cohort data from a folder

Description

loadPlpData loads an object of type plpData from a folder in the file system.

Usage

loadPlpData(file, readOnly = TRUE)

Arguments

file
The name of the folder containing the data.

readOnly
If true, the data is opened read only.

Details

The data will be written to a set of files in the folder specified by the user.

Value

An object of class plpData.

Examples

# todo

[Package PatientLevelPrediction version 5.0.5 Index]