loadPlpFromCsv {PatientLevelPrediction}	R Documentation

Loads parts of the plp result saved as csv files for transparent sharing

Description

Loads parts of the plp result saved as csv files for transparent sharing

Usage

loadPlpFromCsv(dirPath)

Arguments

dirPath
The directory with the results as csv files

Details

Load the main results from csv files into a runPlp object

[Package PatientLevelPrediction version 4.3.10 Index]