loadPredictionAnalysisList {PatientLevelPrediction}	R Documentation

Load the multiple prediction json settings from a file

Description

Load the multiple prediction json settings from a file

Usage

loadPredictionAnalysisList(predictionAnalysisListFile)

Arguments

predictionAnalysisListFile
The prediciton specification json extracted from atlas.

Details

This function interprets a json with the multiple prediction settings and creates a list that can be combined with connection settings to run a multiple prediction study

Examples

## Not run: 
predictionAnalysisList <- loadPredictionAnalysisList('./predictionStudyAnalyses.json')
predictionAnalysisList$connectionDetails = connectionDetails
predictionAnalysisList$cdmDatabaseSchema = cdmDatabaseSchema
predictionAnalysisList$cdmDatabaseName = cdmDatabaseName
predictionAnalysisList$oracleTempSchema = oracleTempSchema
predictionAnalysisList$cohortDatabaseSchema = cohortDatabaseSchema
predictionAnalysisList$cohortTable = cohortTable
predictionAnalysisList$outcomeDatabaseSchema = outcomeDatabaseSchema
predictionAnalysisList$outcomeTable = outcomeTable
predictionAnalysisList$cdmVersion = cdmVersion
predictionAnalysisList$outputFolder = outputFolder
result <- do.call(runPlpAnalyses, predictionAnalysisList)

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]