plotPlp {PatientLevelPrediction}	R Documentation

Plot all the PatientLevelPrediction plots

Description

Plot all the PatientLevelPrediction plots

Usage

plotPlp(plpResult, saveLocation = NULL, typeColumn = "evaluation")

Arguments

plpResult
Object returned by the runPlp() function

saveLocation
Name of the directory where the plots should be saved (NULL means no saving)

typeColumn
The name of the column specifying the evaluation type (to stratify the plots)

Details

Create a directory with all the plots

Value

TRUE if it ran

[Package PatientLevelPrediction version 5.0.5 Index]