plotSmoothCalibration {PatientLevelPrediction}	R Documentation

Plot the smooth calibration as detailed in Calster et al. "A calibration heirarchy for risk models was defined: from utopia to empirical data" (2016)

Description

Plot the smooth calibration as detailed in Calster et al. "A calibration heirarchy for risk models was defined: from utopia to empirical data" (2016)

Usage

plotSmoothCalibration(
  plpResult,
  smooth = "loess",
  span = 0.75,
  nKnots = 5,
  scatter = FALSE,
  bins = 20,
  sample = TRUE,
  typeColumn = "evaluation",
  saveLocation = NULL,
  fileName = "smoothCalibration.pdf"
)

Arguments

plpResult
The result of running runPlp function. An object containing the model or location where the model is save, the data selection settings, the preprocessing and training settings as well as various performance measures obtained by the model.

smooth
options: 'loess' or 'rcs'

span
This specifies the width of span used for loess. This will allow for faster computing and lower memory usage.

nKnots
The number of knots to be used by the rcs evaluation. Default is 5

scatter
plot the decile calibrations as points on the graph. Default is False

bins
The number of bins for the histogram. Default is 20.

sample
If using loess then by default 20,000 patients will be sampled to save time

typeColumn
The name of the column specifying the evaluation type

saveLocation
Directory to save plot (if NULL plot is not saved)

fileName
Name of the file to save to plot, for example 'plot.png'. See the function ggsave in the ggplot2 package for supported file formats.

Details

Create a plot showing the smoothed calibration #'

Value

A ggplot object.

[Package PatientLevelPrediction version 5.0.5 Index]