plotSparseRoc {PatientLevelPrediction}	R Documentation

Plot the ROC curve using the sparse thresholdSummary data frame

Description

Plot the ROC curve using the sparse thresholdSummary data frame

Usage

plotSparseRoc(
  plpResult,
  typeColumn = "evaluation",
  saveLocation = NULL,
  fileName = "roc.png"
)

Arguments

plpResult
A plp result object as generated using the runPlp function.

typeColumn
The name of the column specifying the evaluation type

saveLocation
Directory to save plot (if NULL plot is not saved)

fileName
Name of the file to save to plot, for example 'plot.png'. See the function ggsave in the ggplot2 package for supported file formats.

Details

Create a plot showing the Receiver Operator Characteristics (ROC) curve.

Value

A ggplot object. Use the ggsave function to save to file in a different format.

[Package PatientLevelPrediction version 5.0.5 Index]