predictPlp {PatientLevelPrediction}	R Documentation

predictPlp

Description

Predict the risk of the outcome using the input plpModel for the input plpData

Usage

predictPlp(plpModel, population, plpData, index = NULL)

Arguments

plpModel
An object of type plpModel - a patient level prediction model

population
The population created using createStudyPopulation() who will have their risks predicted

plpData
An object of type plpData - the patient level prediction data extracted from the CDM.

index
A data frame containing rowId: a vector of rowids and index: a vector of doubles the same length as the rowIds. If used, only the rowIds with a negative index value are used to calculate the prediction.

Details

The function applied the trained model on the plpData to make predictions

Value

A dataframe containing the prediction for each person in the population with an attribute metaData containing prediction details.

[Package PatientLevelPrediction version 4.3.10 Index]