recalibratePlp {PatientLevelPrediction}	R Documentation

recalibratePlp

Description

Train various models using a default parameter gird search or user specified parameters

Usage

recalibratePlp(
  prediction,
  analysisId,
  typeColumn = "evaluationType",
  method = c("recalibrationInTheLarge", "weakRecalibration")
)

Arguments

prediction
A prediction dataframe

analysisId
The model analysisId

typeColumn
The column name where the strata types are specified

method
Method used to recalibrate ('recalibrationInTheLarge' or 'weakRecalibration' )

Details

The user can define the machine learning model to train (regularised logistic regression, random forest, gradient boosting machine, neural network and )

Value

An object of class runPlp that is recalibrated on the new data

[Package PatientLevelPrediction version 5.0.5 Index]