recalibratePlpRefit {PatientLevelPrediction}	R Documentation

recalibratePlpRefit

Description

Train various models using a default parameter gird search or user specified parameters

Usage

recalibratePlpRefit(plpModel, newPopulation, newData)

Arguments

plpModel
The trained plpModel (runPlp$model)

newPopulation
The population created using createStudyPopulation() who will have their risks predicted

newData
An object of type plpData - the patient level prediction data extracted from the CDM.

Details

The user can define the machine learning model to train (regularised logistic regression, random forest, gradient boosting machine, neural network and )

Value

An object of class runPlp that is recalibrated on the new data

[Package PatientLevelPrediction version 5.0.5 Index]