savePlpData {PatientLevelPrediction}	R Documentation

Save the cohort data to folder

Description

savePlpData saves an object of type plpData to folder.

Usage

savePlpData(plpData, file, envir = NULL, overwrite = F)

Arguments

plpData
An object of type plpData as generated using getPlpData.

file
The name of the folder where the data will be written. The folder should not yet exist.

envir
The environment for to evaluate variables when saving

overwrite
Whether to force overwrite an existing file

Details

The data will be written to a set of files in the folder specified by the user.

Examples

# todo

[Package PatientLevelPrediction version 5.0.5 Index]