savePlpShareable {PatientLevelPrediction}	R Documentation

Save the plp result as json files and csv files for transparent sharing

Description

Save the plp result as json files and csv files for transparent sharing

Usage

savePlpShareable(result, saveDirectory, minCellCount = 10)

Arguments

result
An object of class runPlp with development or validation results

saveDirectory
The directory the save the results as csv files

minCellCount
Minimum cell count for the covariateSummary and certain evaluation results

Details

Saves the main results json/csv files (these files can be read by the shiny app)

[Package PatientLevelPrediction version 5.0.5 Index]