savePlpToCsv {PatientLevelPrediction}	R Documentation

Save parts of the plp result as a csv for transparent sharing

Description

Save parts of the plp result as a csv for transparent sharing

Usage

savePlpToCsv(result, dirPath)

Arguments

result
An object of class runPlp with development or validation results

dirPath
The directory the save the results as csv files

Details

Saves the main results as a csv (these files can be read by the shiny app)

[Package PatientLevelPrediction version 4.3.10 Index]