Saves the prediction dataframe to a json file

Saves the prediction dataframe to a json file

Usage

savePrediction(prediction, dirPath, fileName = "prediction.json")

Arguments

prediction

The prediciton data.frame

dirPath

The directory to save the prediction json

fileName

The name of the json file that will be saved

Value

                       The file location where the prediction was saved

Details

Saves the prediction data frame returned by predict.R to an json file and returns the fileLocation where the prediction is saved

Examples

prediction <- data.frame(
  rowIds = c(1, 2, 3),
  outcomeCount = c(0, 1, 0),
  value = c(0.1, 0.9, 0.2)
)
saveLoc <- file.path(tempdir())
savePrediction(prediction, saveLoc)
#> [1] "/tmp/Rtmpmk7Npm/prediction.json"
dir(saveLoc)
#>  [1] "bslib-0283b6b1e00c5c093bb7ba6e996bedd7"
#>  [2] "downlit"                               
#>  [3] "file24bf118ceb13.duckdb"               
#>  [4] "file24bf118ceb13.duckdb.wal"           
#>  [5] "file24bf1d9114df"                      
#>  [6] "file24bf1ddc7de8.duckdb"               
#>  [7] "file24bf1ddc7de8.duckdb.wal"           
#>  [8] "file24bf245e7934"                      
#>  [9] "file24bf277d4f3c"                      
#> [10] "file24bf351ecef9"                      
#> [11] "file24bf457a41b1.duckdb"               
#> [12] "file24bf71cc199a.duckdb"               
#> [13] "file24bf71cc199a.duckdb.wal"           
#> [14] "prediction.json"                       
#> [15] "temp_libpath24bf6f84af20"              

# clean up
unlink(file.path(saveLoc, "prediction.json"))