Create setting for CovNN2 model - convolution across input and time - https://arxiv.org/pdf/1608.00647.pdf

Create setting for CovNN2 model - convolution across input and time - https://arxiv.org/pdf/1608.00647.pdf

setCovNN2(batchSize = 1000, outcomeWeight = 1, lr = 1e-05,
  decay = 1e-06, dropout = 0, epochs = 10, filters = 3,
  kernelSize = 10, loss = "binary_crossentropy", seed = NULL)

Arguments

batchSize
The number of samples to used in each batch during model training

outcomeWeight
The weight assined to the outcome (make greater than 1 to reduce unballanced label issue)

lr
The learning rate

decay
The decay of the learning rate

dropout
[currently not used] the dropout rate for regularisation

epochs
The number of times data is used to train the model (e.g., epoches=1 means data only used once to train)

filters
The number of columns output by each convolution

kernelSize
The number of time dimensions used for each convolution

loss
The loss function implemented

seed
The random seed

Examples

# NOT RUN {
model.CovNN <- setCovNN()
# }