setDeepNN {PatientLevelPrediction}	R Documentation

Create setting for DeepNN model

Description

Create setting for DeepNN model

Usage

setDeepNN(
  units = list(c(128, 64), 128),
  layer_dropout = c(0.2),
  lr = c(1e-04),
  decay = c(1e-05),
  outcome_weight = c(1),
  batch_size = c(100),
  epochs = c(100),
  seed = NULL
)

Arguments

units
The number of units of the deep network - as a list of vectors

layer_dropout
The layer dropout rate (regularisation)

lr
Learning rate

decay
Learning rate decay over each update.

outcome_weight
The weight of the outcome class in the loss function

batch_size
The number of data points to use per training batch

epochs
Number of times to iterate over dataset

seed
Random seed used by deep learning model

Examples

## Not run: 
model <- setDeepNN()

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]