Create setting for GBM Survival with python #' @description This creates a setting for fitting GBM surivial model. You need sksurv python install. To install this open your command line and type: conda install -c sebp scikit-survival

Create setting for GBM Survival with python #' @description This creates a setting for fitting GBM surivial model. You need sksurv python install. To install this open your command line and type: conda install -c sebp scikit-survival

setGBMSurvival(
  loss = "coxph",
  learningRate = 0.1,
  nEstimators = c(100),
  criterion = "friedman_mse",
  minSamplesSplit = 2,
  minSamplesLeaf = 1,
  minWeightFractionLeaf = 0,
  maxDepth = c(3, 10, 17),
  minImpuritySplit = NULL,
  minImpurityDecrease = 0,
  maxFeatures = NULL,
  maxLeafNodes = NULL,
  presort = NULL,
  subsample = 1,
  dropoutRate = 0,
  seed = NULL,
  quiet = F
)

Arguments

loss
A string specifying the loss function to minimise (default: 'coxph' )

learningRate
A double specifying the learning rate (controls convergence speed)

nEstimators
An integer specifying how many trees to build

criterion
Default: 'friedman_mse'

minSamplesSplit
An integer specifying min samples per tree split (complexity)

minSamplesLeaf
An integer specifying min samples per leaf (complexity)

minWeightFractionLeaf
Lookup

maxDepth
An integer specifying the max depth of trees (complexity)

minImpuritySplit
A double or NULL specifying the minimum impurity split

minImpurityDecrease
will add

maxFeatures
will add

maxLeafNodes
will add

presort
will add

subsample
will add

dropoutRate
will add

seed
will add

quiet
will add

Details

Pick the hyper-parameters you want to do a grid search for

Examples

if (FALSE) {
gbmSurv <- setGBMSurvival(learningRate=c(0.1,0.01), nEstimators =c(10,50,100),
 maxDepth=c(4,10,17), seed = 2)
}