setKNN {PatientLevelPrediction}	R Documentation

Create setting for knn model

Description

Create setting for knn model

Usage

setKNN(k = 1000, indexFolder = file.path(getwd(), "knn"), threads = 1)

Arguments

k
The number of neighbors to consider

indexFolder
The directory where the results and intermediate steps are output

threads
The number of threads to use when applying big knn

Examples

## Not run: 
model.knn <- setKNN(k=10000)

## End(Not run)

[Package PatientLevelPrediction version 5.0.5 Index]