setLRTorch {PatientLevelPrediction}	R Documentation

Create setting for logistics regression model with python

Description

Create setting for logistics regression model with python

Usage

setLRTorch(
  w_decay = c(5e-04, 0.005),
  epochs = c(20, 50, 100),
  seed = NULL,
  class_weight = 0,
  autoencoder = FALSE,
  vae = FALSE
)

Arguments

w_decay
The l2 regularisation

epochs
The number of epochs

seed
A seed for the model

class_weight
The class weight used for imbalanced data: 0: Inverse ratio between positives and negatives -1: Focal loss

autoencoder
First learn stakced autoencoder for input features, then train LR on the encoded features.

vae
First learn stakced varational autoencoder for input features, then train LR on the encoded features.

Examples

## Not run: 
model.lrTorch <- setLRTorch()

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]