setMLPTorch {PatientLevelPrediction}	R Documentation

Create setting for neural network model with python

Description

Create setting for neural network model with python

Usage

setMLPTorch(
  size = c(500, 1000),
  w_decay = c(5e-04, 0.005),
  epochs = c(20, 50),
  seed = 0,
  class_weight = 0,
  mlp_type = "MLP",
  autoencoder = FALSE,
  vae = FALSE
)

Arguments

size
The number of hidden nodes

w_decay
The l2 regularisation

epochs
The number of epochs

seed
A seed for the model

class_weight
The class weight used for imbalanced data: 0: Inverse ratio between positives and negatives -1: Focal loss

mlp_type
The type of multiple layer network, inlcuding MLP and SNN (self-normalizing neural network)

autoencoder
First learn stakced autoencoder for input features, then train MLP on the encoded features.

vae
First learn stakced varational autoencoder for input features, then train MLP on the encoded features.

Examples

## Not run: 
model.mlpTorch <- setMLPTorch()

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]