simulatePlpData {PatientLevelPrediction}	R Documentation

Generate simulated data

Description

simulateplpData creates a plpData object with simulated data.

Usage

simulatePlpData(plpDataSimulationProfile, n = 10000)

Arguments

plpDataSimulationProfile
An object of type plpDataSimulationProfile as generated using the
createplpDataSimulationProfile function.

n
The size of the population to be generated.

Details

This function generates simulated data that is in many ways similar to the original data on which the simulation profile is based. The contains same outcome, comparator, and outcome concept IDs, and the covariates and their 1st order statistics should be comparable.

Value

An object of type plpData.

[Package PatientLevelPrediction version 5.0.5 Index]