Saves sklearn python model object to json in path

Saves sklearn python model object to json in path

Usage

sklearnToJson(model, path)

Arguments

model

a fitted sklearn python model object

path

path to the saved model file

Value

nothing, saves the model to the path as json

Examples