transportModel {PatientLevelPrediction}	R Documentation

Transports a plpModel to a new location and removes sensitive data

Description

Transports a plpModel to a new location and removes sensitive data

Usage

transportModel(plpModel, outputFolder)

Arguments

plpModel
A trianed model.

outputFolder
The folder on the file system where the CSV files will be created. If the folder does not yet exist it will be created.

Details

This function is used to

Examples

## Not run: 
transportModel(plpModel, "s:/temp/exportTest")

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]