validateMultiplePlp {PatientLevelPrediction}	R Documentation

externally validate the multiple plp models across new datasets

Description

This function loads all the models in a multiple plp analysis folder and validates the models on new data

Usage

validateMultiplePlp(
  analysesLocation,
  validationDatabaseDetails,
  validationRestrictPlpDataSettings = createRestrictPlpDataSettings(),
  recalibrate = NULL,
  saveDirectory = NULL
)

Arguments

analysesLocation
The location where the multiple plp analyses are

validationDatabaseDetails
The validation database settings created using createDatabaseDetails()

validationRestrictPlpDataSettings
The settings specifying the extra restriction settings when extracting the data created using createRestrictPlpDataSettings().

recalibrate
A vector of recalibration methods (currently supports 'RecalibrationintheLarge' and/or 'weakRecalibration')

saveDirectory
The location to save to validation results

Details

Users need to input a location where the results of the multiple plp analyses are found and the connection and database settings for the new data

[Package PatientLevelPrediction version 5.0.5 Index]