viewMultiplePlp {PatientLevelPrediction}	R Documentation

open a local shiny app for viewing the result of a multiple PLP analyses

Description

open a local shiny app for viewing the result of a multiple PLP analyses

Usage

viewMultiplePlp(analysesLocation)

Arguments

analysesLocation
The directory containing the results (with the analysis_x folders)

Details

Opens a shiny app for viewing the results of the models from various T,O, Tar and settings settings.

[Package PatientLevelPrediction version 5.0.5 Index]