viewPlp {PatientLevelPrediction}	R Documentation

viewPlp - Interactively view the performance and model settings

Description

This is a shiny app for viewing interactive plots of the performance and the settings

Usage

viewPlp(runPlp, validatePlp = NULL)

Arguments

runPlp
The output of runPlp() (an object of class 'runPlp')

validatePlp
The output of externalValidatePlp (on object of class 'validatePlp')

Details

Either the result of runPlp and view the plots

Value

Opens a shiny app for interactively viewing the results

[Package PatientLevelPrediction version 5.0.5 Index]