MapIds {PatientLevelPrediction}	R Documentation

Map covariate and row Ids so they start from 1

Description

this functions takes covariate data and a cohort/population and remaps the covariate and row ids, restricts to pop and saves/creates mapping

Usage

MapIds(covariateData, cohort = NULL, mapping = NULL)

Arguments

covariateData
a covariateData object

cohort
if specified rowIds restricted to the ones in cohort

mapping
A pre defined mapping to use

[Package PatientLevelPrediction version 6.3.1 Index]