Populate the PatientLevelPrediction results tables

This function formats and uploads results that have been generated via an ATLAS prediction package into a database

Usage

addMultipleRunPlpToDatabase(
  connectionDetails,
  databaseSchemaSettings = createDatabaseSchemaSettings(resultSchema = "main"),
  cohortDefinitions,
  databaseList = NULL,
  resultLocation = NULL,
  resultLocationVector,
  modelSaveLocation
)

Arguments

connectionDetails

A connection details created by using the function createConnectionDetails in the DatabaseConnector package.

databaseSchemaSettings

A object created by createDatabaseSchemaSettings with all the settings specifying the result tables

cohortDefinitions

A set of one or more cohorts extracted using ROhdsiWebApi::exportCohortDefinitionSet()

databaseList

(Optional) A list created by createDatabaseList to specify the databases

resultLocation

(string) location of directory where the main package results were saved

resultLocationVector

(only used when resultLocation is missing) a vector of locations with development or validation results

modelSaveLocation

The location of the file system for saving the models in a subdirectory

Value

Returns NULL but uploads all the results in resultLocation to the PatientLevelPrediction result tables in resultSchema

Details

This function can be used upload PatientLevelPrediction results into a database