configurePython {PatientLevelPrediction}	R Documentation

Sets up a virtual environment to use for PLP (can be conda or python)

Description

Sets up a virtual environment to use for PLP (can be conda or python)

Usage

configurePython(envname = "PLP", envtype = NULL)

Arguments

envname
A string for the name of the virtual environment (default is 'PLP')

envtype
An option for specifying the environment as'conda' or 'python'. If NULL then the default is 'conda' for windows users and 'python' for non-windows users

Details

This function creates a virtual environment that can be used by PatientLevelPrediction and installs all the required package dependancies. If using python, pip must be set up.

[Package PatientLevelPrediction version 5.0.5 Index]