createPlpResultTables {PatientLevelPrediction}	R Documentation

Create the results tables to store PatientLevelPrediction models and results into a database

Description

This function executes a large set of SQL statements to create tables that can store models and results

Usage

createPlpResultTables(
  conn,
  resultSchema,
  targetDialect = "postgresql",
  deleteExistingTables = T,
  createTables = T,
  stringAppendToTables = "",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  testFile = NULL
)

Arguments

conn
A connection to a database created by using the function connect in the DatabaseConnector package.

resultSchema
The name of the database schema that the result tables will be created.

targetDialect
The database management system being used

deleteExistingTables
If true any existing tables matching the PatientLevelPrediction result tables names will be deleted

createTables
If true the PatientLevelPrediction result tables will be created

stringAppendToTables
A string that appends to the PatientLevelPrediction result tables

tempEmulationSchema
The temp schema used when the database management system is oracle

testFile
(used for testing) The location of an sql file with the table creation code

Details

This function can be used to create (or delete) PatientLevelPrediction result tables

Value

Returns NULL but creates the required tables into the specified database schema.

[Package PatientLevelPrediction version 5.0.5 Index]