Create a study population

Create a study population

createStudyPopulation(plpData, population = NULL, outcomeId, binary = T,
  includeAllOutcomes = T, firstExposureOnly = FALSE, washoutPeriod = 0,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  requireTimeAtRisk = T, minTimeAtRisk = 365, riskWindowStart = 0,
  addExposureDaysToStart = FALSE, riskWindowEnd = 365,
  addExposureDaysToEnd = F, verbosity = "INFO", ...)

Arguments

plpData
An object of type plpData as generated using getDbplpData.

population
If specified, this population will be used as the starting point instead of the cohorts in the plpData object.

outcomeId
The ID of the outcome. If not specified, no outcome-specific transformations will be performed.

binary
Forces the outcomeCount to be 0 or 1 (use for binary prediction problems)

includeAllOutcomes
(binary) indicating whether to include people with outcomes who are not observed for the whole at risk period

firstExposureOnly
Should only the first exposure per subject be included? Note that this is typically done in the createStudyPopulation function,

washoutPeriod
The mininum required continuous observation time prior to index date for a person to be included in the cohort.

removeSubjectsWithPriorOutcome
Remove subjects that have the outcome prior to the risk window start?

priorOutcomeLookback
How many days should we look back when identifying prior outcomes?

requireTimeAtRisk
Should subject without time at risk be removed?

minTimeAtRisk
The minimum number of days at risk required to be included

riskWindowStart
The start of the risk window (in days) relative to the index date (+ days of exposure if the addExposureDaysToStart parameter is specified).

addExposureDaysToStart
Add the length of exposure the start of the risk window?

riskWindowEnd
The end of the risk window (in days) relative to the index data (+ days of exposure if the addExposureDaysToEnd parameter is specified).

addExposureDaysToEnd
Add the length of exposure the risk window?

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

...
Other inputs

Value

A data frame specifying the study population. This data frame will have the following columns:

rowId

A unique identifier for an exposure

subjectId

The person ID of the subject

cohortStartdate

The index date

outcomeCount

The number of outcomes observed during the risk window

timeAtRisk

The number of days in the risk window

survivalTime

The number of days until either the outcome or the end of the risk window

Details

Create a study population by enforcing certain inclusion and exclusion criteria, defining a risk window, and determining which outcomes fall inside the risk window.