createStudyPopulation {PatientLevelPrediction}	R Documentation

Create a study population

Description

Create a study population

Usage

createStudyPopulation(
  plpData,
  population = NULL,
  outcomeId,
  binary = T,
  includeAllOutcomes = T,
  firstExposureOnly = FALSE,
  washoutPeriod = 0,
  removeSubjectsWithPriorOutcome = TRUE,
  priorOutcomeLookback = 99999,
  requireTimeAtRisk = F,
  minTimeAtRisk = 365,
  riskWindowStart = 0,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  verbosity = "INFO",
  restrictTarToCohortEnd = F,
  addExposureDaysToStart,
  addExposureDaysToEnd,
  ...
)

Arguments

plpData
An object of type plpData as generated using getplpData.

population
If specified, this population will be used as the starting point instead of the cohorts in the plpData object.

outcomeId
The ID of the outcome.

binary
Forces the outcomeCount to be 0 or 1 (use for binary prediction problems)

includeAllOutcomes
(binary) indicating whether to include people with outcomes who are not observed for the whole at risk period

firstExposureOnly
Should only the first exposure per subject be included? Note that this is typically done in the createStudyPopulation function,

washoutPeriod
The mininum required continuous observation time prior to index date for a person to be included in the cohort.

removeSubjectsWithPriorOutcome
Remove subjects that have the outcome prior to the risk window start?

priorOutcomeLookback
How many days should we look back when identifying prior outcomes?

requireTimeAtRisk
Should subject without time at risk be removed?

minTimeAtRisk
The minimum number of days at risk required to be included

riskWindowStart
The start of the risk window (in days) relative to the startAnchor.

startAnchor
The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".

riskWindowEnd
The end of the risk window (in days) relative to the endAnchor parameter

endAnchor
The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

restrictTarToCohortEnd
If using a survival model and you want the time-at-risk to end at the cohort end date set this to T

addExposureDaysToStart
DEPRECATED: Add the length of exposure the start of the risk window? Use startAnchor instead.

addExposureDaysToEnd
DEPRECATED: Add the length of exposure the risk window? Use endAnchor instead.

...
Other inputs

Details

Create a study population by enforcing certain inclusion and exclusion criteria, defining a risk window, and determining which outcomes fall inside the risk window.

Value

A data frame specifying the study population. This data frame will have the following columns:

rowId

A unique identifier for an exposure

subjectId

The person ID of the subject

cohortStartdate

The index date

outcomeCount

The number of outcomes observed during the risk window

timeAtRisk

The number of days in the risk window

survivalTime

The number of days until either the outcome or the end of the risk window

[Package PatientLevelPrediction version 4.3.10 Index]