diagnostic {PatientLevelPrediction}	R Documentation

diagnostic - Investigates the prediction problem settings - use before training a model

Description

This function runs a set of prediction diagnoses to help pick a suitable T, O, TAR and determine whether the prediction problem is worth executing.

Usage

diagnostic(
  plpData = NULL,
  cdmDatabaseName = "none",
  cohortName,
  outcomeNames,
  databaseDetails,
  restrictPlpDataSettings,
  populationSettings,
  outputFolder = NULL,
  minCellCount = 5
)

Arguments

plpData
The data object to do the diagnostic on - if NULL you need to specify the connection settings below

cdmDatabaseName
The name of the database being diagnosed

cohortName
Name of the target cohort

outcomeNames
Vector of outcome names

databaseDetails
(only used is plpData is NULL) The database details created using createDatabaseDetails

restrictPlpDataSettings
(only used is plpData is NULL) The restrictPlpDataSettings created using createRestrictPlpDataSettings

populationSettings
The population setting details created using createPopulationSettings

outputFolder
Location to save results for shiny app

minCellCount
The minimum count that will be displayed

Details

Users can define set of Ts, Os, databases and population settings. A list of data.frames containing details such as follow-up time distribution, time-to-event information, characteriszation details, time from last prior event, observation time distribution.

Value

An object containing the model or location where the model is save, the data selection settings, the preprocessing and training settings as well as various performance measures obtained by the model.

distribution
list for each O of a data.frame containing: i) Time to observation end distribution, ii) Time from observation start distribution, iii) Time to event distribution and iv) Time from last prior event to index distribution (only for patients in T who have O before index)

incident
list for each O of incidence of O in T during TAR

characterization
list for each O of Characterization of T, TnO, Tn~O

Examples

## Not run: 
#******** EXAMPLE 1 ********* 

## End(Not run) 

[Package PatientLevelPrediction version 5.0.5 Index]