evaluateMultiplePlp {PatientLevelPrediction}	R Documentation

externally validate the multiple plp models across new datasets

Description

This function loads all the models in a multiple plp analysis folder and validates the models on new data

Usage

evaluateMultiplePlp(
  analysesLocation,
  outputLocation,
  connectionDetails,
  validationSchemaTarget,
  validationSchemaOutcome,
  validationSchemaCdm,
  databaseNames,
  validationTableTarget,
  validationTableOutcome,
  validationIdTarget = NULL,
  validationIdOutcome = NULL,
  oracleTempSchema = NULL,
  verbosity = "INFO",
  keepPrediction = F,
  recalibrate = NULL,
  sampleSize = NULL
)

Arguments

analysesLocation
The location where the multiple plp analyses are

outputLocation
The location to save to validation results

connectionDetails
The connection details for extracting the new data

validationSchemaTarget
A string or list of strings specifying the database containing the target cohorts

validationSchemaOutcome
A string or list of strings specifying the database containing the outcome cohorts

validationSchemaCdm
A string or list of strings specifying the database containing the cdm

databaseNames
A string of lift of strings specifying sharing friendly database names corresponding to validationSchemaCdm

validationTableTarget
A string or list of strings specifying the table containing the target cohorts

validationTableOutcome
A string or list of strings specifying the table containing the outcome cohorts

validationIdTarget
An iteger or list of integers specifying the cohort id for the target cohorts

validationIdOutcome
An iteger or list of integers specifying the cohort id for the outcome cohorts

oracleTempSchema
The temp oracle schema requires read/write

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

keepPrediction
Whether to keep the predicitons for the new data

recalibrate
A vector of recalibration methods (currently supports 'RecalibrationintheLarge' and/or 'weakRecalibration')

sampleSize
If not NULL, the number of people to sample from the target cohort

Details

Users need to input a location where the results of the multiple plp analyses are found and the connection and database settings for the new data

[Package PatientLevelPrediction version 4.3.10 Index]