fitPlp {PatientLevelPrediction}	R Documentation

fitPlp

Description

Train various models using a default parameter gird search or user specified parameters

Usage

fitPlp(trainData, modelSettings, search = "grid", analysisId)

Arguments

trainData
An object of type TrainData created using splitData data extracted from the CDM.

modelSettings
An object of class modelSettings created using one of the function:

logisticRegressionModel() A lasso logistic regression model

GBMclassifier() A gradient boosting machine

RFclassifier() A random forest model

GLMclassifier () A generalised linear model

KNNclassifier() A KNN model

search
The search strategy for the hyper-parameter selection (currently not used)

analysisId
The id of the analysis

Details

The user can define the machine learning model to train (regularised logistic regression, random forest, gradient boosting machine, neural network and )

Value

An object of class plpModel containing:

model
The trained prediction model

preprocessing
The preprocessing required when applying the model

prediction
The cohort data.frame with the predicted risk column added

modelDesign
A list specifiying the modelDesign settings used to fit the model

trainDetails
The model meta data

covariateImportance
The covariate importance for the model

[Package PatientLevelPrediction version 6.3.1 Index]