fitPlp {PatientLevelPrediction}	R Documentation

fitPlp

Description

Train various models using a default parameter gird search or user specified parameters

Usage

fitPlp(
  population,
  data,
  modelSettings,
  cohortId,
  outcomeId,
  minCovariateFraction = 0.001,
  normalizeData = T
)

Arguments

population
The population created using createStudyPopulation() who will have their risks predicted

data
An object of type plpData - the patient level prediction data extracted from the CDM.

modelSettings
An object of class modelSettings created using one of the function:

logisticRegressionModel() A lasso logistic regression model

GBMclassifier() A gradient boosting machine

RFclassifier() A random forest model

GLMclassifier () A generalised linear model

KNNclassifier() A KNN model

cohortId
Id of study cohort

outcomeId
Id of outcome cohort

minCovariateFraction
The minimum fraction of the target popualtion who have a variable for it to be included in the model training

normalizeData
Whether to normalise the data before model fitting

Details

The user can define the machine learning model to train (regularised logistic regression, random forest, gradient boosting machine, neural network and )

Value

An object of class plpModel containing:

model
The trained prediction model

modelLoc
The path to where the model is saved (if saved)

trainAuc
The AUC obtained on the training set

trainCalibration
The calibration obtained on the training set

modelSettings
A list specifiying the model, preprocessing, outcomeId and cohortId

metaData
The model meta data

trainingTime
The time taken to train the classifier

[Package PatientLevelPrediction version 4.3.10 Index]