Get a sparse summary of the calibration

Get a sparse summary of the calibration

Usage

getCalibrationSummary(
  prediction,
  predictionType,
  typeColumn = "evaluation",
  numberOfStrata = 10,
  truncateFraction = 0.05
)

Arguments

prediction

A prediction object as generated using the predict functions.

predictionType

The type of prediction (binary or survival)

typeColumn

A column that is used to stratify the results

numberOfStrata

The number of strata in the plot.

truncateFraction

This fraction of probability values will be ignored when plotting, to avoid the x-axis scale being dominated by a few outliers.

Value

A dataframe with the calibration summary

Details

Generates a sparse summary showing the predicted probabilities and the observed fractions. Predictions are stratified into equally sized bins of predicted probabilities.

Examples

# simulate data
data("simulationProfile")
plpData <- simulatePlpData(simulationProfile, n=500)
#> Generating covariates
#> Generating cohorts
#> Generating outcomes
# create study population, split into train/test and preprocess with default settings
population <- createStudyPopulation(plpData, outcomeId = 3)
#> outcomeId: 3
#> binary: TRUE
#> includeAllOutcomes: TRUE
#> firstExposureOnly: FALSE
#> washoutPeriod: 0
#> removeSubjectsWithPriorOutcome: TRUE
#> priorOutcomeLookback: 99999
#> requireTimeAtRisk: TRUE
#> minTimeAtRisk: 364
#> restrictTarToCohortEnd: FALSE
#> riskWindowStart: 1
#> startAnchor: cohort start
#> riskWindowEnd: 365
#> endAnchor: cohort start
#> restrictTarToCohortEnd: FALSE
#> Removing subjects with prior outcomes (if any)
#> Removing non outcome subjects with insufficient time at risk (if any)
#> Outcome is 0 or 1
#> Population created with: 481 observations, 481 unique subjects and 231 outcomes
#> Population created in 0.0395 secs
data <- splitData(plpData, population, createDefaultSplitSetting())
#> test: 0.25
#> train: 0.75
#> nfold: 3
#> seed: 2020
#> Creating a 25% test and 75% train (into 3 folds) random stratified split by class
#> Data split into 119 test cases and 362 train cases (121, 121, 120)
#> Starting to limit covariate data to population...
#> Finished limiting covariate data to population...
#> Starting to limit covariate data to population...
#> Finished limiting covariate data to population...
#> Data split in 0.233 secs
data$Train$covariateData <- preprocessData(data$Train$covariateData)
#> minFraction: 0.001
#> normalize: TRUE
#> removeRedundancy: TRUE
#> Removing 1 redundant covariates
#> Removing 0 infrequent covariates
#> Normalizing covariates
#> Tidying covariates took 0.672 secs
saveLoc <- file.path(tempdir(), "calibrationSummary")
# fit a lasso logistic regression model using the training data
plpModel <- fitPlp(data$Train, modelSettings=setLassoLogisticRegression(seed=42),
                   analysisId=1, analysisPath=saveLoc)
#> Running Cyclops
#> Done.
#> GLM fit status:  OK
#> Returned from fitting to LassoLogisticRegression
#> Getting variable importance
#> Creating variable importance data frame
#> Getting predictions on train set
#> predictProbabilities - predictAndromeda start
#> Prediction took 0.104 secs
#> Returned from classifier function
#> Time to fit model: 0.22 secs
calibrationSummary <- getCalibrationSummary(plpModel$prediction, 
                                            "binary", 
                                            numberOfStrata = 10,
                                            typeColumn = "evaluationType")
calibrationSummary
#>    predictionThreshold PersonCountAtRisk PersonCountWithOutcome
#> 1            0.0000000                72                     28
#> 2            0.3967779                 4                      0
#> 3            0.4182838                85                     34
#> 4            0.4311129                70                     31
#> 5            0.4746809                71                     40
#> 6            0.5100570                31                     19
#> 7            0.6209443                29                     22
#> 8            0.0000000                51                     24
#> 9            0.3953152                24                     14
#> 10           0.4136041                36                     14
#> 11           0.4389705                37                     17
#> 12           0.4443968                38                     22
#> 13           0.4698104                33                     10
#> 14           0.4816173                51                     24
#> 15           0.5028033                20                      9
#> 16           0.5371844                36                     16
#> 17           0.5973235                36                     24
#>    averagePredictedProbability StDevPredictedProbability
#> 1                    0.3935215               0.015228158
#> 2                    0.4182838               0.000000000
#> 3                    0.4311129               0.000000000
#> 4                    0.4746809               0.000000000
#> 5                    0.5098516               0.001730388
#> 6                    0.5965674               0.024012417
#> 7                    0.6699312               0.014078745
#> 8                    0.3874111               0.008500192
#> 9                    0.4064082               0.003716602
#> 10                   0.4346486               0.006989355
#> 11                   0.4419036               0.002741549
#> 12                   0.4662758               0.007653728
#> 13                   0.4788592               0.004643898
#> 14                   0.4968033               0.005687440
#> 15                   0.5261732               0.007376023
#> 16                   0.5670503               0.022460997
#> 17                   0.6449116               0.025450689
#>    MinPredictedProbability P25PredictedProbability MedianPredictedProbability
#> 1                0.3123910               0.3967779                  0.3967779
#> 2                0.4182838               0.4182838                  0.4182838
#> 3                0.4311129               0.4311129                  0.4311129
#> 4                0.4746809               0.4746809                  0.4746809
#> 5                0.4954765               0.5100570                  0.5100570
#> 6                0.5308364               0.5870932                  0.5870932
#> 7                0.6613927               0.6613927                  0.6613927
#> 8                0.3546579               0.3837830                  0.3843002
#> 9                0.3968635               0.4053957                  0.4053957
#> 10               0.4219420               0.4357438                  0.4389705
#> 11               0.4389705               0.4389705                  0.4443968
#> 12               0.4458033               0.4698104                  0.4698104
#> 13               0.4698104               0.4728699                  0.4816173
#> 14               0.4875726               0.4923061                  0.4955607
#> 15               0.5140176               0.5227849                  0.5263849
#> 16               0.5434985               0.5434985                  0.5754406
#> 17               0.6104719               0.6227571                  0.6507883
#>    P75PredictedProbability MaxPredictedProbability observedIncidence evaluation
#> 1                0.3967779               0.3967779         0.3888889      Train
#> 2                0.4182838               0.4182838         0.0000000      Train
#> 3                0.4311129               0.4311129         0.4000000      Train
#> 4                0.4746809               0.4746809         0.4428571      Train
#> 5                0.5100570               0.5100570         0.5633803      Train
#> 6                0.6209443               0.6209443         0.6129032      Train
#> 7                0.6923446               0.6923446         0.7586207      Train
#> 8                0.3953152               0.3953152         0.4705882         CV
#> 9                0.4053957               0.4136041         0.5833333         CV
#> 10               0.4389705               0.4389705         0.3888889         CV
#> 11               0.4443968               0.4443968         0.4594595         CV
#> 12               0.4698104               0.4698104         0.5789474         CV
#> 13               0.4816173               0.4816173         0.3030303         CV
#> 14               0.5028033               0.5028033         0.4705882         CV
#> 15               0.5347731               0.5371844         0.4500000         CV
#> 16               0.5852895               0.5973235         0.4444444         CV
#> 17               0.6552410               0.7147491         0.6666667         CV
# clean up
unlink(saveLoc, recursive = TRUE)