Get a sparse summary of the calibration

Get a sparse summary of the calibration

Usage

getCalibrationSummary(
  prediction,
  predictionType,
  typeColumn = "evaluation",
  numberOfStrata = 10,
  truncateFraction = 0.05
)

Arguments

prediction

A prediction object as generated using the predict functions.

predictionType

The type of prediction (binary or survival)

typeColumn

A column that is used to stratify the results

numberOfStrata

The number of strata in the plot.

truncateFraction

This fraction of probability values will be ignored when plotting, to avoid the x-axis scale being dominated by a few outliers.

Value

A dataframe with the calibration summary

Details

Generates a sparse summary showing the predicted probabilities and the observed fractions. Predictions are stratified into equally sized bins of predicted probabilities.

Examples

# simulate data
data("simulationProfile")
plpData <- simulatePlpData(simulationProfile, n=500)
#> Generating covariates
#> Generating cohorts
#> Generating outcomes
# create study population, split into train/test and preprocess with default settings
population <- createStudyPopulation(plpData, outcomeId = 3)
#> outcomeId: 3
#> binary: TRUE
#> includeAllOutcomes: TRUE
#> firstExposureOnly: FALSE
#> washoutPeriod: 0
#> removeSubjectsWithPriorOutcome: TRUE
#> priorOutcomeLookback: 99999
#> requireTimeAtRisk: TRUE
#> minTimeAtRisk: 364
#> restrictTarToCohortEnd: FALSE
#> riskWindowStart: 1
#> startAnchor: cohort start
#> riskWindowEnd: 365
#> endAnchor: cohort start
#> restrictTarToCohortEnd: FALSE
#> Removing subjects with prior outcomes (if any)
#> Removing non outcome subjects with insufficient time at risk (if any)
#> Outcome is 0 or 1
#> Population created with: 483 observations, 483 unique subjects and 233 outcomes
#> Population created in 0.0414 secs
data <- splitData(plpData, population, createDefaultSplitSetting())
#> test: 0.25
#> train: 0.75
#> nfold: 3
#> seed: 75775
#> Creating a 25% test and 75% train (into 3 folds) random stratified split by class
#> Data split into 120 test cases and 363 train cases (122, 121, 120)
#> Starting to limit covariate data to population...
#> Finished limiting covariate data to population...
#> Starting to limit covariate data to population...
#> Finished limiting covariate data to population...
#> Data split in 0.763 secs
data$Train$covariateData <- preprocessData(data$Train$covariateData)
#> minFraction: 0.001
#> normalize: TRUE
#> removeRedundancy: TRUE
#> Removing 1 redundant covariates
#> Removing 0 infrequent covariates
#> Normalizing covariates
#> Tidying covariates took 1.2 secs
saveLoc <- file.path(tempdir(), "calibrationSummary")
# fit a lasso logistic regression model using the training data
plpModel <- fitPlp(data$Train, modelSettings=setLassoLogisticRegression(seed=42),
                   analysisId=1, analysisPath=saveLoc)
#> Running Cyclops
#> Done.
#> GLM fit status:  OK
#> Returned from fitting to LassoLogisticRegression
#> Getting variable importance
#> Creating variable importance data frame
#> Getting predictions on train set
#> predictProbabilities - predictAndromeda start
#> Prediction took 0.162 secs
#> Returned from classifier function
#> Time to fit model: 0.34 secs
calibrationSummary <- getCalibrationSummary(plpModel$prediction, 
                                            "binary", 
                                            numberOfStrata = 10,
                                            typeColumn = "evaluationType")
calibrationSummary
#>    predictionThreshold PersonCountAtRisk PersonCountWithOutcome
#> 1            0.0000000                86                     29
#> 2            0.3862219                29                     12
#> 3            0.4227336                92                     39
#> 4            0.4397387                11                      8
#> 5            0.4478777                37                     16
#> 6            0.4989161                35                     21
#> 7            0.6214513                36                     23
#> 8            0.6708076                37                     27
#> 9            0.0000000                38                     16
#> 10           0.3964319                36                     17
#> 11           0.4190337                38                     14
#> 12           0.4208268                42                     14
#> 13           0.4357011                28                     15
#> 14           0.4397177                43                     21
#> 15           0.4632520                29                     12
#> 16           0.5078695                42                     23
#> 17           0.5949383                30                     19
#> 18           0.6530045                37                     24
#>    averagePredictedProbability StDevPredictedProbability
#> 1                    0.3826749              0.0067767442
#> 2                    0.4162459              0.0116983791
#> 3                    0.4386274              0.0036754045
#> 4                    0.4446562              0.0003721349
#> 5                    0.4737807              0.0125073555
#> 6                    0.5452762              0.0424326447
#> 7                    0.6312294              0.0150942234
#> 8                    0.6874332              0.0212373606
#> 9                    0.3737215              0.0187817683
#> 10                   0.4079824              0.0038437379
#> 11                   0.4201410              0.0008707856
#> 12                   0.4298384              0.0063933570
#> 13                   0.4368809              0.0016166432
#> 14                   0.4569184              0.0057233368
#> 15                   0.4758155              0.0145058836
#> 16                   0.5567212              0.0363445510
#> 17                   0.6147343              0.0199305037
#> 18                   0.7008847              0.0324304289
#>    MinPredictedProbability P25PredictedProbability MedianPredictedProbability
#> 1                0.3622672               0.3862219                  0.3862219
#> 2                0.3911199               0.4147074                  0.4227336
#> 3                0.4251824               0.4397387                  0.4397387
#> 4                0.4439036               0.4448235                  0.4448235
#> 5                0.4600944               0.4610800                  0.4773596
#> 6                0.5014228               0.5162451                  0.5177583
#> 7                0.6226150               0.6226150                  0.6226150
#> 8                0.6729727               0.6729727                  0.6729727
#> 9                0.3300787               0.3646785                  0.3646785
#> 10               0.4042652               0.4050912                  0.4086875
#> 11               0.4190337               0.4190337                  0.4208268
#> 12               0.4208268               0.4248095                  0.4329971
#> 13               0.4357011               0.4357011                  0.4357011
#> 14               0.4414145               0.4542988                  0.4542988
#> 15               0.4648032               0.4648032                  0.4648032
#> 16               0.5078695               0.5182265                  0.5768796
#> 17               0.5976164               0.6017476                  0.6094904
#> 18               0.6530746               0.6692046                  0.6936688
#>    P75PredictedProbability MaxPredictedProbability observedIncidence evaluation
#> 1                0.3862219               0.3862219         0.3372093      Train
#> 2                0.4227336               0.4227336         0.4137931      Train
#> 3                0.4397387               0.4397387         0.4239130      Train
#> 4                0.4448235               0.4448235         0.7272727      Train
#> 5                0.4798769               0.4989161         0.4324324      Train
#> 6                0.5982903               0.6197058         0.6000000      Train
#> 7                0.6310196               0.6621469         0.6388889      Train
#> 8                0.7054237               0.7367001         0.7297297      Train
#> 9                0.3964319               0.3964319         0.4210526         CV
#> 10               0.4086875               0.4190337         0.4722222         CV
#> 11               0.4208268               0.4208268         0.3684211         CV
#> 12               0.4357011               0.4357011         0.3333333         CV
#> 13               0.4383906               0.4397177         0.5357143         CV
#> 14               0.4632520               0.4632520         0.4883721         CV
#> 15               0.4845111               0.5078695         0.4137931         CV
#> 16               0.5949383               0.5949383         0.5476190         CV
#> 17               0.6094904               0.6527245         0.6333333         CV
#> 18               0.7398389               0.7616838         0.6486486         CV
# clean up
unlink(saveLoc, recursive = TRUE)