getCalibrationSummary {PatientLevelPrediction}	R Documentation

Get a sparse summary of the calibration

Description

Get a sparse summary of the calibration

Usage

getCalibrationSummary(
  prediction,
  predictionType,
  typeColumn = "evaluation",
  numberOfStrata = 100,
  truncateFraction = 0.05
)

Arguments

prediction
A prediction object as generated using the predict functions.

predictionType
The type of prediction (binary or survival)

typeColumn
A column that is used to stratify the results

numberOfStrata
The number of strata in the plot.

truncateFraction
This fraction of probability values will be ignored when plotting, to avoid the x-axis scale being dominated by a few outliers.

Details

Generates a sparse summary showing the predicted probabilities and the observed fractions. Predictions are stratefied into equally sized bins of predicted probabilities.

Value

A dataframe with the calibration summary

[Package PatientLevelPrediction version 5.0.5 Index]