getPlpTable {PatientLevelPrediction}	R Documentation

Create a dataframe with the summary details of the population cohort for publications

Description

Create a dataframe with the summary details of the population cohort for publications

Usage

getPlpTable(
  cdmDatabaseSchema,
  oracleTempSchema,
  covariateSettings,
  longTermStartDays = -365,
  population,
  connectionDetails,
  cohortTable = "#temp_person"
)

Arguments

cdmDatabaseSchema
The schema containing the OMOP CDM data

oracleTempSchema
The oracle schema if needed

covariateSettings
The covariateSettings if different from default

longTermStartDays
How far to look back when looking for the variables in the data

population
The population you want the summary table for

connectionDetails
The connection details used to connect to the CDM database

cohortTable
The name of the temp table that will store the popualtion cohort

Details

This function is used to create a summary table for population to be inserted into publications

Examples

## Not run: 
getTable1 (plpData, population, connectionDetails)

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]