listCartesian {PatientLevelPrediction}	R Documentation

Cartesian product

Description

Computes the Cartesian product of all the combinations of elements in a list

Usage

listCartesian(allList)

Arguments

allList
a list of lists

Value

A list with all possible combinations from the input list of lists

[Package PatientLevelPrediction version 6.3.1 Index]