preprocessData {PatientLevelPrediction}	R Documentation

A function that wraps around FeatureExtraction::tidyCovariateData to normalise the data and remove rare or redundant features

Description

A function that wraps around FeatureExtraction::tidyCovariateData to normalise the data and remove rare or redundant features

Usage

preprocessData(covariateData, preprocessSettings)

Arguments

covariateData
The covariate part of the training data created by splitData after being sampled and having any required feature engineering

preprocessSettings
The settings for the preprocessing created by createPreprocessSettings

Details

Returns an object of class covariateData that has been processed

Value

The data processed

[Package PatientLevelPrediction version 5.0.5 Index]