runMultiplePlp {PatientLevelPrediction}	R Documentation

Run a list of predictions analyses

Description

Run a list of predictions analyses

Usage

runMultiplePlp(
  databaseDetails = createDatabaseDetails(),
  modelDesignList = list(createModelDesign(targetId = 1, outcomeId = 2, modelSettings =
    setLassoLogisticRegression()), createModelDesign(targetId = 1, outcomeId = 3,
    modelSettings = setLassoLogisticRegression())),
  onlyFetchData = F,
  splitSettings = createDefaultSplitSetting(type = "stratified", testFraction = 0.25,
    trainFraction = 0.75, splitSeed = 123, nfold = 3),
  cohortDefinitions = NULL,
  logSettings = createLogSettings(verbosity = "DEBUG", timeStamp = T, logName =
    "runPlp Log"),
  saveDirectory = getwd()
)

Arguments

databaseDetails
The database settings created using createDatabaseDetails()

modelDesignList
A list of model designs created using createModelDesign()

onlyFetchData
Only fetches and saves the data object to the output folder without running the analysis.

splitSettings
The train/validation/test splitting used by all analyses created using createDefaultSplitSetting()

cohortDefinitions
A list of cohort definitions for the target and outcome cohorts

logSettings
The setting spexcifying the logging for the analyses created using createLogSettings()

saveDirectory
Name of the folder where all the outputs will written to.

Details

This function will run all specified predictions as defined using .

Value

A data frame with the following columns:

analysisId 	 The unique identifier for a set of analysis choices.

cohortId 	 The ID of the target cohort populations.

outcomeId 	 The ID of the outcomeId.

dataLocation 	 The location where the plpData was saved

evaluationFolder 	 The name of file containing the evaluation saved as a csv

the settings ids 	 The ids for all other settings used for model development.

[Package PatientLevelPrediction version 5.0.5 Index]