setAdaBoost {PatientLevelPrediction}	R Documentation

Create setting for AdaBoost with python

Description

Create setting for AdaBoost with python

Usage

setAdaBoost(nEstimators = 50, learningRate = 1, seed = NULL)

Arguments

nEstimators
The maximum number of estimators at which boosting is terminated

learningRate
Learning rate shrinks the contribution of each classifier by learningRate. There is a trade-off between learningRate and nEstimators .

seed
A seed for the model

Examples

## Not run: 
model.adaBoost <- setAdaBoost(size = 4, alpha = 1e-05, seed = NULL)

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]