setCoxModel {PatientLevelPrediction}	R Documentation

Create setting for lasso Cox model

Description

Create setting for lasso Cox model

Usage

setCoxModel(variance = 0.01, seed = NULL)

Arguments

variance
a single value used as the starting value for the automatic lambda search

seed
An option to add a seed when training the model

Examples

model.lr <- setCoxModel()

[Package PatientLevelPrediction version 4.3.10 Index]