setGradientBoostingMachine {PatientLevelPrediction}	R Documentation

Create setting for gradient boosting machine model using gbm_xgboost implementation

Description

Create setting for gradient boosting machine model using gbm_xgboost implementation

Usage

setGradientBoostingMachine(
  ntrees = c(100, 1000),
  nthread = 20,
  earlyStopRound = 25,
  maxDepth = c(4, 6, 17),
  minRows = 2,
  learnRate = c(0.005, 0.01, 0.1),
  seed = NULL
)

Arguments

ntrees
The number of trees to build

nthread
The number of computer threads to (how many cores do you have?)

earlyStopRound
If the performance does not increase over earlyStopRound number of interactions then training stops (this prevents overfitting)

maxDepth
Maximum number of interactions - a large value will lead to slow model training

minRows
The minimum number of rows required at each end node of the tree

learnRate
The boosting learn rate

seed
An option to add a seed when training the final model

Examples

model.gbm <- setGradientBoostingMachine(ntrees=c(10,100), nthread=20,
                           maxDepth=c(4,6), learnRate=c(0.1,0.3))

[Package PatientLevelPrediction version 4.3.10 Index]