setLassoLogisticRegression {PatientLevelPrediction}	R Documentation

Create setting for lasso logistic regression

Description

Create setting for lasso logistic regression

Usage

setLassoLogisticRegression(
  variance = 0.01,
  seed = NULL,
  includeCovariateIds = c(),
  noShrinkage = c(0),
  threads = -1,
  useCrossValidation = TRUE,
  upperLimit = 20,
  lowerLimit = 0.01
)

Arguments

variance
a single value used as the starting value for the automatic lambda search

seed
An option to add a seed when training the model

includeCovariateIds
a set of covariate IDS to limit the analysis to

noShrinkage
a set of covariates whcih are to be forced to be included in the final model. default is the intercept

threads
An option to set number of threads when training model

useCrossValidation
Set this to FALSE if you want to train a LR with a preset varience

upperLimit
Upper prior variance limit for grid-search

lowerLimit
Lower prior variance limit for grid-search

Examples

model.lr <- setLassoLogisticRegression()

[Package PatientLevelPrediction version 4.3.10 Index]