setMLP {PatientLevelPrediction}	R Documentation

Create setting for neural network model with python

Description

Create setting for neural network model with python

Usage

setMLP(
  size = 4,
  alpha = c(0.3, 0.01, 0.001, 1e-06),
  maxIter = 2000,
  tol = 1e-04,
  learningRateInit = 0.001,
  nIterNoChange = 10,
  beta1 = 0.9,
  beta2 = 0.999,
  epsilon = c(1, 0.1, 1e-08),
  seed = NULL
)

Arguments

size
The number of hidden nodes

alpha
The l2 regularisation

maxIter
Maximum number of iterations. The solver iterates until convergence (determined by ‘tol’) or this number of iterations.

tol
Tolerance for the optimization

learningRateInit
The initial learning rate used. It controls the step-size in updating the weights.

nIterNoChange
Maximum number of epochs to not meet tol improvement.

beta1
Exponential decay rate for estimates of first moment vector in adam, should be in [0, 1).

beta2
Exponential decay rate for estimates of second moment vector in adam, should be in [0, 1).

epsilon
Value for numerical stability in adam.

seed
A seed for the model

Examples

## Not run: 
model.mlp <- setMLP(size=4, alpha=0.00001, seed=NULL)

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]