setRandomForest {PatientLevelPrediction}	R Documentation

Create setting for random forest model with python (very fast)

Description

Create setting for random forest model with python (very fast)

Usage

setRandomForest(
  mtries = -1,
  ntrees = 500,
  maxDepth = c(4, 10, 17),
  varImp = T,
  seed = NULL
)

Arguments

mtries
The number of features to include in each tree (-1 defaults to square root of total features)

ntrees
The number of trees to build

maxDepth
Maximum number of interactions - a large value will lead to slow model training

varImp
Perform an initial variable selection prior to fitting the model to select the useful variables

seed
An option to add a seed when training the final model

Examples

## Not run: 
model.rf <- setRandomForest(mtries=c(-1,5,20),  ntrees=c(10,100), 
                           maxDepth=c(5,20))

## End(Not run)                           

[Package PatientLevelPrediction version 4.3.10 Index]