transportPlp {PatientLevelPrediction}	R Documentation

Transports a plpResult to a new location and removed sensitive data

Description

Transports a plpResult to a new location and removed sensitive data

Usage

transportPlp(
  plpResult,
  modelName = NULL,
  dataName = NULL,
  outputFolder,
  n = NULL,
  includeEvaluationStatistics = T,
  includeThresholdSummary = T,
  includeDemographicSummary = T,
  includeCalibrationSummary = T,
  includePredictionDistribution = T,
  includeCovariateSummary = T,
  save = T,
  reduceSize = F
)

Arguments

plpResult
An object returned by running runPlp.

modelName
A string of the name of the model

dataName
A string of the name of the data

outputFolder
The folder on the file system where the CSV files will be created. If the folder does not yet exist it will be created.

n
The minimum number of people required for each result summary to be included

includeEvaluationStatistics
Whether to include the evaluationStatistics

includeThresholdSummary
Whether to include the thresholdSummary

includeDemographicSummary
Whether to include the demographicSummary

includeCalibrationSummary
Whether to include the calibrationSummary

includePredictionDistribution
Whether to include the predictionDistribution

includeCovariateSummary
Whether to include the covariateSummary

save
Whether to save the result or just return the transportable object

reduceSize
Remove parts of runPlp object that are not needed but take up space

Details

This function is used to

Examples

## Not run: 
transportPlp(plpResult, "s:/temp/exportTest", n=10)

## End(Not run)

[Package PatientLevelPrediction version 4.3.10 Index]